scripts/sde/requirements.txt
# SDE extraction script dependencies
PyYAML>=6.0
